requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
altgraph==0.17.2
anyio==3.6.2
appdirs==1.4.4
APScheduler==3.6.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.5.2
asttokens==2.2.1
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
audioread==3.0.0
autopep8==1.6.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==6.0.0
bs4==0.0.1
cachetools==4.2.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
comm==0.1.2
cryptography==38.0.1
cssselect==1.1.0
cssutils==2.5.0
dateparser==1.1.1
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
demjson3==3.0.5
distlib==0.3.5
Django==4.1.2
django-filter==22.1
django-ratelimit==4.0.0
django-reviews==1.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
easy-thumbnails==2.8.5
executing==1.2.0
facebook-scraper==0.2.57
fake-useragent==0.1.11
fastjsonschema==2.16.2
filelock==3.8.0
Flask==2.1.3
fqdn==1.5.1
frozenlist==1.3.0
future==0.18.2
h11==0.13.0
idna==3.4
importlib-metadata==4.12.0
ipykernel==6.21.2
ipython==8.10.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.5.1
jupyter-events==0.6.3
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyter_server==2.3.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
keyboard==0.13.5
librosa==0.9.2
llvmlite==0.39.1
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==2.0.5
multidict==6.0.2
nbclassic==0.5.1
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.2
outcome==1.2.0
packaging==23.0
pandocfilters==1.5.0
parse==1.19.0
parso==0.8.3
pefile==2022.5.30
pickleshare==0.7.5
pika==1.3.1
Pillow==9.2.0
pip-review==1.3.0
platformdirs==2.5.2
pooch==1.6.0
premailer==3.10.0
prometheus-client==0.16.0
prompt-toolkit==3.0.36
psutil==5.9.4
pure-eval==0.2.2
pycodestyle==2.8.0
pycparser==2.21
pydub==0.25.1
pyee==8.2.2
Pygments==2.14.0
pyinstaller==5.2
pyinstaller-hooks-contrib==2022.8
PyJWT==2.6.0
pyOpenSSL==22.0.0
pyppeteer==1.0.2
pyquery==1.4.3
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.19.0
python-json-logger==2.0.6
python-telegram-bot==13.13
python3-openid==3.2.0
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
pywin32==304
pywin32-ctypes==0.2.0
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
regex==2022.3.2
requests==2.28.1
requests-html==0.10.0
requests-oauthlib==1.3.1
resampy==0.4.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
scikit-learn==1.2.1
scipy==1.10.0
selenium==4.3.0
Send2Trash==1.8.0
six==1.16.0
slackclient==2.9.4
sniffio==1.2.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.6.2
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
toml==0.10.2
tornado==6.2
tqdm==4.64.0
traitlets==5.9.0
trio==0.21.0
trio-websocket==0.9.2
tzdata==2022.1
tzlocal==4.2
uri-template==1.2.0
urllib3==1.26.12
virtualenv==20.16.3
w3lib==1.22.0
wcwidth==0.2.6
webcolors==1.12
webdriver-manager==3.8.1
webencodings==0.5.1
websocket-client==1.5.1
websockets==10.3
Werkzeug==2.1.2
widgetsnbextension==4.0.5
wsproto==1.1.0
yagmail==0.15.280
yarl==1.7.2
zipp==3.8.1
